backend/requirements.txt
fastapi
uvicorn
pandas
requests
fyers-apiv3
python-dotenv
pydantic
pydantic-settings
